+++ src/main.rs
@@ -3,12 +3,22 @@ mod event;
 mod pain;
 mod renderer;
 
-use std::fs::read_to_string;
-
 use crate::editor::Editor;
 use crate::renderer::Renderer;
 
-pub use crate::editor::document::Spot;
+use std::fs::read_to_string;
+
+#[derive(Eq, PartialEq, Clone, Copy)]
+pub struct Spot {
+    pub x: usize,
+    pub y: usize,
+}
+
+impl Spot {
+    pub fn is_zero(&self) -> bool {
+        return self.x == 0 && self.y == 0;
+    }
+}
 
 fn main() -> Result<(), std::io::Error> {
     let source = if let Some(path) = std::env::args().nth(1) {
